requirements.txt
dropbox==3.42
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
onedrivesdk==1.0.1
PyYAML==3.11
requests==2.8.1
six==1.10.0
urllib3==1.12
Werkzeug==0.11.2
wheel==0.24.0
